rename juridische-houdbaarheid-tooltip
renamed file and all files importing the component
