+++ pages/houtbouw/gemeentelijke-omgevingsvisie.js
@@ -3,7 +3,7 @@ import Image from 'next/image';
 import Layout from '../../components/layout';
 import IcontWood from '../../public/icons/houtbouwIconBg.svg';
 import RTooltip from '../../components/r_ladder_tooltip';
-import JHTooltip from '../../components/juridische_houdbaarheid_tooltip';
+import JHTooltip from '../../components/juridische-houdbaarheid-tooltip';
 import JITooltip from '../../components/juridische_invloed_tooltip';
 
 function classNames(...classes) {
