+++ pages/measures/[id].jsx
@@ -6,7 +6,7 @@ import Layout from '../../components/layout';
 import WindmillIcon from '../../public/winturbines.svg';
 import IcontWood from '../../public/icons/houtbouwIconBg.svg';
 import RTooltip from '../../components/r_ladder_tooltip';
-import JHTooltip from '../../components/juridische_houdbaarheid_tooltip';
+import JHTooltip from '../../components/juridische-houdbaarheid-tooltip';
 import JITooltip from '../../components/juridische_invloed_tooltip';
 
 const formatDate = (date) => {
